Feature/be/s12 p21 c107 209 카드추천 로직

See merge request s12-fintech-finance-sub1/S12P21C107!240

+++ BE/src/main/java/com/c107/recommendcard/entity/RecommendCard.java
@@ -17,7 +17,7 @@ public class RecommendCard {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "reco_card_id")
-    private Integer recoCardId;
+    private Long recoCardId;
 
     @Column(name = "reco_card_name")
     private String recoCardName;
@@ -39,4 +39,5 @@ public class RecommendCard {
 
     @Column(name = "source_card_id")
     private Integer sourceCardId;
+
 }
